feat: Criação de pastas DDD e exceptionhandler
Foi criado as pastas domain e api para organização e com a implementação do handler foi removido vários try catch do PessoaController
