+++ fiquepositivo/src/main/java/br/com/fiquepositivo/api/exceptionhandler/ApiExceptionHandler.java
@@ -1,0 +1,22 @@
+package br.com.fiquepositivo.api.exceptionhandler;
+
+import br.com.fiquepositivo.domain.exceptions.ConflitoDeDadosException;
+import br.com.fiquepositivo.domain.exceptions.IdNaoCadastradoException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ApiExceptionHandler {
+
+    @ExceptionHandler(IdNaoCadastradoException.class)
+    public ResponseEntity<String> TratarIdNaoCadastradoException(IdNaoCadastradoException e) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    }
+
+    @ExceptionHandler(ConflitoDeDadosException.class)
+    public ResponseEntity<String> TratarConflitoDeDadosException(ConflitoDeDadosException e) {
+        return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
+    }
+}
