+++ fiquepositivo/src/main/java/br/com/fiquepositivo/domain/model/Pessoa.java
@@ -1,48 +1,35 @@
-package br.com.fiquepositivo.model;
-
-import java.time.LocalDate;
+package br.com.fiquepositivo.domain.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
-public class Gasto {
+public class Pessoa {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@EqualsAndHashCode.Include
 	private Integer id;
 	
+	@Column(nullable = false, length = 60)
+	private String nome;
+	
 	@Column(nullable = false)
-	private Double valor;
-
-	@Column
-	private FormaPagamentoENUM formaPagamento;
-
-	@Column
-	private LocalDate data;
+	private Double RendaMensal;
 	
 	@Column(length = 60)
-	private String local;
-
-	@Column
-	private String descricao;
-	
-	@Column
-	private Boolean necessidade;
-	
-	@JoinColumn(nullable = false)
-	@ManyToOne
-	private Pessoa pessoa;
+	private String profissao;
 	
 	
 }
