+++ fiquepositivo/src/main/java/br/com/fiquepositivo/domain/repository/GastoRepository.java
@@ -1,9 +1,9 @@
-package br.com.fiquepositivo.repository;
+package br.com.fiquepositivo.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.fiquepositivo.model.Gasto;
+import br.com.fiquepositivo.domain.model.Gasto;
 
 @Repository
 public interface GastoRepository extends JpaRepository<Gasto, Integer> {
