+++ fiquepositivo/src/main/java/br/com/fiquepositivo/domain/repository/PessoaRepository.java
@@ -1,9 +1,9 @@
-package br.com.fiquepositivo.repository;
+package br.com.fiquepositivo.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.fiquepositivo.model.Pessoa;
+import br.com.fiquepositivo.domain.model.Pessoa;
 
 @Repository
 public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
